% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbListTables.R
\name{dbListTables,TrinoConnection-method}
\alias{dbListTables,TrinoConnection-method}
\title{List the tables in the connection's schema}
\usage{
\S4method{dbListTables}{TrinoConnection}(conn, ...)
}
\arguments{
\item{conn}{A \link[=TrinoConnection-class]{TrinoConnection} object.}

\item{...}{Unused, for compatibility with the generic.}
}
\value{
A character vector of table names.
}
\description{
List the tables in the connection's schema
}
\examples{
\dontshow{if (nzchar(Sys.getenv("RTRINO_TEST_URL"))) withAutoprint(\{ # examplesIf}
con <- DBI::dbConnect(RTrino::Trino(), host = Sys.getenv("RTRINO_TEST_URL"),
                      catalog = "tpch", schema = "tiny")
DBI::dbListTables(con)
DBI::dbDisconnect(con)
\dontshow{\}) # examplesIf}
}
