% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbRemoveTable.R
\name{dbRemoveTable,TrinoConnection,character-method}
\alias{dbRemoveTable,TrinoConnection,character-method}
\alias{dbRemoveTable,TrinoConnection,Id-method}
\alias{dbRemoveTable,TrinoConnection,ANY-method}
\title{Drop a table}
\usage{
\S4method{dbRemoveTable}{TrinoConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{TrinoConnection,Id}(conn, name, ...)

\S4method{dbRemoveTable}{TrinoConnection,ANY}(conn, name, ...)
}
\arguments{
\item{conn}{A \link[=TrinoConnection-class]{TrinoConnection} object.}

\item{name}{Table name, in any form \code{\link[DBI:dbExistsTable]{dbExistsTable()}} accepts.}

\item{...}{Unused, for compatibility with the generic.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Drop a table
}
\examples{
\dontrun{
DBI::dbRemoveTable(con, "sales")
}
}
