% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinoDriver.R
\name{Trino}
\alias{Trino}
\title{Instantiate the Trino driver}
\usage{
Trino()
}
\value{
A \link[=TrinoDriver-class]{TrinoDriver} object.
}
\description{
Creates the driver object passed as the first argument to
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}. The call has no side effects and always succeeds.
}
\examples{
drv <- Trino()
drv
}
