% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrinoDriver.R
\docType{class}
\name{TrinoDriver-class}
\alias{TrinoDriver-class}
\alias{show,TrinoDriver-method}
\alias{dbIsValid,TrinoDriver-method}
\alias{dbGetInfo,TrinoDriver-method}
\alias{dbUnloadDriver,TrinoDriver-method}
\title{Trino driver class}
\usage{
\S4method{show}{TrinoDriver}(object)

\S4method{dbIsValid}{TrinoDriver}(dbObj, ...)

\S4method{dbGetInfo}{TrinoDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{TrinoDriver}(drv, ...)
}
\arguments{
\item{object}{A \link[=TrinoDriver-class]{TrinoDriver} object.}

\item{dbObj}{A \link[=TrinoDriver-class]{TrinoDriver} object.}

\item{...}{Unused, for compatibility with the generic.}

\item{drv}{A \link[=TrinoDriver-class]{TrinoDriver} object.}
}
\description{
An S4 class representing the Trino driver. It carries no state and exists
only as the entry point for \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.
}
\keyword{internal}
