% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbExistsTable.R
\name{dbExistsTable,TrinoConnection,character-method}
\alias{dbExistsTable,TrinoConnection,character-method}
\title{Does a table exist?}
\usage{
\S4method{dbExistsTable}{TrinoConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A \link[=TrinoConnection-class]{TrinoConnection} object.}

\item{name}{Table name, as a string or an identifier.}

\item{...}{Unused, for compatibility with the generic.}
}
\value{
A logical scalar.
}
\description{
Does a table exist?
}
\examples{
\dontrun{
DBI::dbExistsTable(con, "sales")
}
}
