% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbListTables.R
\name{dbListTables,TrinoConnection-method}
\alias{dbListTables,TrinoConnection-method}
\title{List the tables in the connection's schema}
\usage{
\S4method{dbListTables}{TrinoConnection}(conn, ...)
}
\arguments{
\item{conn}{A \link[=TrinoConnection-class]{TrinoConnection} object.}

\item{...}{Unused, for compatibility with the generic.}
}
\value{
A character vector of table names.
}
\description{
List the tables in the connection's schema
}
\examples{
\dontrun{
DBI::dbListTables(con)
}
}
