% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTrino-package.R
\docType{package}
\name{RTrino-package}
\alias{RTrino}
\alias{RTrino-package}
\title{RTrino: 'DBI' Backend for 'Trino'}
\description{
A 'DBI' compliant backend for 'Trino' \url{https://trino.io}, the distributed SQL query engine (successor of 'PrestoSQL'). Connects to a 'Trino' cluster over its HTTP REST API, executes queries with automatic pagination, maps 'Trino' types to R types, and integrates with 'dplyr' through 'dbplyr' SQL dialects. Supports basic, JSON Web Token (JWT) bearer and OAuth2 client credentials authentication, as well as custom certificate authorities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AngelCasasbl/RTrino}
  \item Report bugs at \url{https://github.com/AngelCasasbl/RTrino/issues}
}

}
\author{
\strong{Maintainer}: Luis A Casas Ballestas \email{luisangelcasasbl@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Luis A Casas Ballestas \email{luisangelcasasbl@gmail.com} [copyright holder]
}

}
\keyword{internal}
